add timeout handler for batch loader
https://github.com/graph-gophers/graphql-go allows configuration of max
parallelism which defines the maximum number of resolvers running at a
time. With parallelism set low, or many resolvers running the dataloader
can block callers (waiting for more keys to be added an capacity to be
hit). The timeout value allows the batch load function to be executed
after a specific time interval in order to prevent blocking resolver go
routines.