Update strategy and dataloader to return callback functions (Thunk)
Instead of blocking and returning the requested data, Load and LoadMany
now return a callback function (Thunk or ThunkMany respectively). This
allows the ability to call Load and increment the loads count without
blocking allowing the caller to perform some other long running
operations before calling the Thunk function in order to get back the
result (or block if the batch function hasn't ran yet).

The Thunk and ThunkMany functions are linked to the provided keys they
were returned for. Calling the functions will block the caller until the
batch function is called and data is available