Fix race condition and ensure only 1 worker is started
Ensure that only one worker is started and running when callers and
batch functions take a variable amount of time to execute. This
addresses an issue where from time to time a worker would be started
with and the batch function would be called with null keys even though
the previous worker resolved the value for the key that started the
second worker.

Previous fix caused a race condition, this fix ensures that no race
condition is present and unnecessary workers are not started.

more: https://stackoverflow.com/a/11121616/2605221