Update ResultMap interface to return the result and bool
ResultMap's GetValue should return the result for a specific key and a
boolean specifying if the value was found. This ensures that keys
skipped by the batch function and nil data returned from the database
can be easily identified. Previous version would return a Result with a
nil value regardless if the key existed or if the value was found and
nil

[Fixes #1]